added delete buttons

+++ app/Http/Controllers/PostController.php
@@ -20,21 +20,21 @@ class PostController extends Controller
         $filter = $request->input("filter");
         $page = $request->input("page");
         $perPage = $request->input("perPage");
-        $toSkip = ($page-1) * $perPage;
+        $toSkip = ($page - 1) * $perPage;
         // $posts = Post::where('title','LIKE','%'.$filter.'%')
         //                 ->orWhere('body','LIKE','%'.$filter.'%')
         //                 ->orderBy($field, $order)->skip($toSkip)->take($perPage)->get();
         $posts = Post::name($filter)
-                        ->body($filter)
-                        ->title($filter)
-                        ->order($field, $order)
-                        ->skip($toSkip)
-                        ->take($perPage)
-                        ->get();
+            ->body($filter)
+            ->title($filter)
+            ->order($field, $order)
+            ->skip($toSkip)
+            ->take($perPage)
+            ->get();
         $countAndPosts = json_encode(array($count, $posts));
         return response($countAndPosts, 200)
-                ->header('X-Total-Count', $count)
-                ->header('Access-Control-Expose-Headers', 'X-Total-Count');
+            ->header('X-Total-Count', $count)
+            ->header('Access-Control-Expose-Headers', 'X-Total-Count');
     }
 
     /**
@@ -104,7 +104,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->delete();
+        if ($post) {
+            $post->delete();
+        } else {
+            return response()->json(array(), 200);
+        }
         return response()->json(null, 204);
     }
 
@@ -113,8 +117,8 @@ class PostController extends Controller
         $filter = $request->input("filter");
         $filter = json_decode($filter);
         $ids = trim(json_encode($filter->id), "[]"); // $ids is of the string type
-        $postsToDelete = Post::whereIn('id',explode(",",$ids))->get();
-        Post::whereIn('id',explode(",",$ids))->delete();
+        $postsToDelete = Post::whereIn('id', explode(",", $ids))->get();
+        Post::whereIn('id', explode(",", $ids))->delete();
         return response()->json($postsToDelete, 200);
     }
 }
